refactor(core): centralize protocol constants and types
This change consolidates protocol-related constants and types from key-package-relay-list.ts and key-package.ts into a new protocol.ts module. The index.ts file has been updated to export the centralized module, improving code organization and maintainability.

Also ran format
