+++ src/core/index.ts
@@ -2,3 +2,4 @@ export * from "./credential.js";
 export * from "./key-package-relay-list.js";
 export * from "./key-package-store.js";
 export * from "./key-package.js";
+export * from "./protocol.js";
