+++ src/core/key-package.ts
@@ -4,33 +4,15 @@ import { CiphersuiteId, ciphersuites } from "ts-mls/crypto/ciphersuite.js";
 import { decodeKeyPackage } from "ts-mls/keyPackage.js";
 import { getTagValue, NostrEvent } from "../utils/nostr.js";
 import { isValidRelayUrl, normalizeRelayUrl } from "../utils/relay-url.js";
-
-/** Event kind for key package events */
-export const KEY_PACKAGE_KIND = 443;
-
-/** The name of the tag that contains the MLS protocol version */
-export const KEY_PACKAGE_MLS_VERSION_TAG = "mls_protocol_version";
-
-/** The name of the tag that contains the MLS cipher suite */
-export const KEY_PACKAGE_CIPHER_SUITE_TAG = "mls_ciphersuite";
-
-/** The name of the tag that contains the MLS extensions */
-export const KEY_PACKAGE_EXTENSIONS_TAG = "mls_extensions";
-
-/** The name of the tag that contains the relays */
-export const KEY_PACKAGE_RELAYS_TAG = "relays";
-
-/** The name of the tag that contains the client */
-export const KEY_PACKAGE_CLIENT_TAG = "client";
-
-/** The possible MLS protocol versions */
-export type MLS_VERSIONS = "1.0";
-
-/** Parsed client tag from a kind 443 event */
-export type KeyPackageClient = {
-  name: string;
-  // TODO: this is probably a NIP-89 client tag, so it should probably have the rest of the fields
-};
+import {
+  KEY_PACKAGE_CIPHER_SUITE_TAG,
+  KEY_PACKAGE_CLIENT_TAG,
+  KEY_PACKAGE_EXTENSIONS_TAG,
+  KEY_PACKAGE_MLS_VERSION_TAG,
+  KEY_PACKAGE_RELAYS_TAG,
+  KeyPackageClient,
+  MLS_VERSIONS,
+} from "./protocol.js";
 
 /** Get the {@link KeyPackage} from a kind 443 event */
 export function getKeyPackage(event: NostrEvent): KeyPackage {
