+++ src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./nostr.js";
 export * from "./relay-url.js";
-
